fixed limelight shuffleboard

+++ src/main/java/frc/robot/Constants.java
@@ -60,8 +60,8 @@ public class Constants {
 	//number of degrees from perfectly vertical
 	public static final double TARGET_HIEGHT  = 1;
 	//height from floor to target in inches
-	public static final double SHORT_PIPE_NUM = 0;
-	public static final double TALL_PIPE_NUM  = 1;
+	public static final int SHORT_PIPE_NUM = 0;
+	public static final int TALL_PIPE_NUM  = 1;
 
 	Constants() {
